% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.cmai}
\alias{coef.cmai}
\title{S3 class to Estimated regression coefficients for cmai models}
\usage{
\method{coef}{cmai}(object, ...)
}
\arguments{
\item{object}{an object of the class cmai}

\item{...}{further arguments passed to or from other methods}
}
\value{
a list containing the the estimated regression coefficients associated with the degenerated and count distributions, respectively.
}
\description{
S3 class to Estimated regression coefficients for cmai models
}
