% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.cms}
\alias{coef.cms}
\title{S3 class to Estimated regression coefficients for countmodeg model}
\usage{
\method{coef}{cms}(object, ...)
}
\arguments{
\item{object}{an object of the class cms}

\item{...}{further arguments passed to or from other methods}
}
\value{
a vector with the estimated regression coefficients.
}
\description{
S3 class to Estimated regression coefficients for countmodeg model
}
