% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmodel.R
\name{countmodel}
\alias{countmodel}
\title{Regression for count models}
\usage{
countmodel(
  formula,
  family = c("poisson", "negbinom"),
  case = c("standard", "inflated", "hurdle"),
  data = NULL,
  approach = c("mle", "bayes"),
  hessian = TRUE,
  link1 = c("log", "sqrt", "identity"),
  link2 = c("logit", "probit", "cloglog", "cauchy"),
  hyperpars = list(mu_psi = NULL, sigma_psi = NULL, mu_beta = NULL, sigma_beta = NULL,
    a_theta = NULL, b_theta = NULL),
  ...
)
}
\arguments{
\item{formula}{an object of class "formula" (or one that can be coerced to
that class): a symbolic description of the model to be fitted.}

\item{family}{family distribution for the regression model (poisson, negbinom)
; default is poisson}

\item{case}{case of the regression (standard, inflated, hurdle); default is
standard}

\item{data}{data an optional data.frame, list or enviroment containing the
variables in the model}

\item{approach}{approach to be used to fit the model (mle: maximum likelihood;
bayes : Bayesian approach); default is mle}

\item{hessian}{logical ; if TRUE (default), the hessian matrix is returned
when approach = "mle"}

\item{link1}{assumed link function for count distribution (log, sqrt or
identiy); default is log.}

\item{link2}{assumed link function for degenerate distribution (logit, probit,
cloglog, cauchy); default is logit.}

\item{...}{further arguments}

\item{hyperparsa}{list containing the hyperparameters associated with the
prior distribution of the regression coefficients;
if not specified then default choice is an minor informative prior}
}
\value{
an object of class (cms,cmai,cmps,cmnb,cmzips,cmzaps,cmzinb,cmzanb)
depending on the family and case used
}
\description{
Fits an regression model for poisson and  Negative Binomial an their inflated
and hurdle versions
}
