% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.cms}
\alias{fitted.cms}
\title{Predictions for mean for each observation}
\usage{
\method{fitted}{cms}(object, ...)
}
\arguments{
\item{object}{an object of the class cms}

\item{...}{further arguments passed to or from other methods}
}
\value{
vector of the prediction for each observation
}
\description{
Predictions for mean for each observation
}
