% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cms}
\alias{predict.cms}
\title{Predict and Confidence intervals of mean for newdata}
\usage{
\method{predict}{cms}(
  object,
  newdata,
  level = 0.95,
  type = c("link", "response", "terms"),
  ...
)
}
\arguments{
\item{object}{an object of the class cms}

\item{newdata}{an matrix or data.frame with all the covariables used in the model
referring to the new observations}

\item{level}{the confidence level required}

\item{type}{of the confidence interval ("link", "response", "terms");
default is type = "link" **(in development, for now forces use of lik intervals)}

\item{...}{further arguments passed to or from other methods}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence
limits for each observation.
}
\description{
Predict and Confidence intervals of mean for newdata
}
