% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cmnb.R
\name{print.summary.cmnb}
\alias{print.summary.cmnb}
\title{Print the summary.cmnb output}
\usage{
\method{print}{summary.cmnb}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.cmnb.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.cmnb output
}
