% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cmps.R
\name{print.summary.cmps}
\alias{print.summary.cmps}
\title{Print the summary.cmps output}
\usage{
\method{print}{summary.cmps}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.cmps.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.cmps output
}
