% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_negbinom.R
\name{print.summary.cmzanb}
\alias{print.summary.cmzanb}
\title{Print the summary.cmzanb output}
\usage{
\method{print}{summary.cmzanb}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.cmzanb.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.cmzanb output
}
