% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_negbinom.R
\name{print.summary.cmzinb}
\alias{print.summary.cmzinb}
\title{Print the summary.cmzinb output}
\usage{
\method{print}{summary.cmzinb}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.cmzinb.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.cmzinb output
}
