% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cmzips.R
\name{print.summary.cmzips}
\alias{print.summary.cmzips}
\title{Print the summary.cmzips output}
\usage{
\method{print}{summary.cmzips}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.cmzips.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.cmzips output
}
