% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_zipoisreg.R
\name{print.summary.zipoisreg}
\alias{print.summary.zipoisreg}
\title{Print the summary.zipoisreg output}
\usage{
\method{print}{summary.zipoisreg}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.zipoisreg.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.zipoisreg output
}
