% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{residuals.cmai}
\alias{residuals.cmai}
\title{Residuals for mean for each observation}
\usage{
\method{residuals}{cmai}(object, ...)
}
\arguments{
\item{object}{an object of the class cms}

\item{...}{further arguments passed to or from other methods}
}
\value{
vector of the prediction for each observation
}
\description{
Residuals for mean for each observation
}
