% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cmnb.R
\name{summary.cmnb}
\alias{summary.cmnb}
\title{Summary for the standard poisson  model}
\usage{
\method{summary}{cmnb}(object, ...)
}
\arguments{
\item{object}{an objecto of the class 'cmnb'.}

\item{...}{further arguments passed to or from other methods}
}
\description{
Summary for the standard poisson  model
}
