% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cmps.R
\name{summary.cmps}
\alias{summary.cmps}
\title{Summary for the standard poisson  model}
\usage{
\method{summary}{cmps}(object, ...)
}
\arguments{
\item{object}{an objecto of the class 'cmps'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary for the standard poisson  model
}
