% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_negbinom.R
\name{summary.cmzanb}
\alias{summary.cmzanb}
\title{Summary for the zero adjusted negative binomial model}
\usage{
\method{summary}{cmzanb}(object, ...)
}
\arguments{
\item{object}{an objecto of the class 'cmzanb'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary for the zero adjusted negative binomial model
}
