% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cmzips.R
\name{summary.cmzaps}
\alias{summary.cmzaps}
\title{Summary for the zero adjusted poisson model}
\usage{
\method{summary}{cmzaps}(object, ...)
}
\arguments{
\item{object}{an objecto of the class 'cmzaps'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary for the zero adjusted poisson model
}
