% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_negbinom.R
\name{summary.cmzinb}
\alias{summary.cmzinb}
\title{Summary for the zero inflated negative binomial model}
\usage{
\method{summary}{cmzinb}(object, ...)
}
\arguments{
\item{object}{an objecto of the class 'cmzinb'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary for the zero inflated negative binomial model
}
