% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.cmai}
\alias{vcov.cmai}
\title{Variance-covariance matrix for a inflated or hurdle count model}
\usage{
\method{vcov}{cmai}(object, ...)
}
\arguments{
\item{object}{an object of the class cmai.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the variance-covariance matrix associated with the regression coefficients.
}
\description{
This function extracts and returns the variance-covariance matrix associated with the regression coefficients when the maximum likelihood estimation approach is used in the model fitting.
}
