% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countmodels-package.R
\docType{package}
\name{countmodels-package}
\alias{countmodels-package}
\alias{countmodels}
\title{The 'countmodels' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.36.0.9000. https://mc-stan.org
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
