% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cmzips.R
\name{print.summary.cmzaps}
\alias{print.summary.cmzaps}
\title{Print the summary.cmzaps output}
\usage{
\method{print}{summary.cmzaps}(x, ...)
}
\arguments{
\item{x}{an object of the class summary.cmzaps.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a summary of the fitted model.
}
\description{
Print the summary.cmzaps output
}
