% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_negbinom.R
\name{summary.cmzinb}
\alias{summary.cmzinb}
\alias{summary.zipoisreg}
\title{Summary for the zipoisreg model}
\usage{
\method{summary}{cmzinb}(object, ...)
}
\arguments{
\item{object}{an objecto of the class 'zipoisreg'.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary for the zipoisreg model
}
